Refactor: Change menu and create smaller methods to simplify code
  Broke up some of the scraper methods to not be all together, so we 
have more smaller methods that do specific things like scrape_title and 
scrape_url that make it a little easier to read the code.  The 
anti-duplication code to the Job instance creating code in Scaper class 
so we have unique job instances and if we scrape more pages it will 
start scraping x number of pages from the last page we scraped 
succesfully.   Also added some aeshetics to the menu and progress bar.